third_party/blink/web_tests/installedapp/getinstalledrelatedapps-empty.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/installedapp/installed_app_provider.mojom.js"></script>
<script src="resources/installedapp-test-helper.js"></script>
<script>

installedapp_test((t, mock) => {
  // With no manifest, expect the Mojo call to FilterInstalledApps to be empty,
  // and return an empty list.
  mock.pushExpectedCall([], []);
  return navigator.getInstalledRelatedApps().then(result => {
    assert_array_relatedapplication_equals(result, []);
  });
}, 'getInstalledRelatedApps with no apps');

</script>
